third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-contain/container-queries/font-relative-units-expected.txt

This is a testharness.js-based test.
[FAIL] rex relative inline-size
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
[FAIL] rcap relative inline-size
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
